+++ src/components/leftContent/PlatformContent.tsx
